Finally the app is working!
With a call to the server and receiving words from it.

what was the error?
NetworkOnMainThreadException as it is always good to use Asynctask to perform network operations
Secondly, the syntax!
"select ********* where level='"+strings[0]+"'";
I thought the single quote is needed. But no. It takes the parameter as 'easy'

So skip out on the single quotes X

The app is still slow and needs to be fixed.
Wondering how!